% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{dbCreatePK}
\alias{dbCreatePK}
\title{dbCreatePK}
\usage{
dbCreatePK(dbcon, table_name, pk_fields)
}
\arguments{
\item{dbcon}{...}

\item{table_name}{...}

\item{pk_fields}{...}
}
\description{
dbCreatePK
}
