% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{dbExecFile}
\alias{dbExecFile}
\title{dbExecFile}
\usage{
dbExecFile(input_file, dbcon, plist = NULL)
}
\arguments{
\item{input_file}{the file name (including path) to be read.}

\item{dbcon}{database connection, as created by the dbConnect function.}

\item{plist}{a list with value to be binded to parameters of
SQL statements}
}
\description{
dbExecFile
}
