% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromCSV.R
\name{dbTableFromCSV}
\alias{dbTableFromCSV}
\title{The dbTableFromCSV function reads the data from a rectangular region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If the destination table does not exist, it will create it.}
\usage{
dbTableFromCSV(
  input_file,
  schema_file,
  dbcon,
  table_name,
  header = FALSE,
  drop_table = FALSE,
  auto_pk = FALSE,
  build_pk = FALSE,
  constant_values = NULL,
  chunk_size = 10000,
  ...
)
}
\arguments{
\item{input_file}{the file name (including path) to be read}

\item{schema_file}{...}

\item{dbcon}{...}

\item{table_name}{...}

\item{header}{...}

\item{drop_table}{...}

\item{auto_pk}{...}

\item{build_pk}{...}

\item{constant_values}{...}

\item{chunk_size}{...}

\item{...}{...}
}
\value{
nothing
}
\description{
The dbTableFromCSV function reads the data from a rectangular region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If the destination table does not exist, it will create it.
}
