% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromFeather.R
\name{dbTableFromFeather}
\alias{dbTableFromFeather}
\title{The dbTableFromFeather function reads the data from a Apache
Arrow table serialized in a Feather (Arrow IPC) file and copies it
to a table in a SQLite database. If table does not exist, it will
create it.}
\usage{
dbTableFromFeather(
  input_file,
  dbcon,
  table_name,
  columns = NULL,
  drop_table = FALSE,
  auto_pk = FALSE,
  pk_fields = NULL,
  build_pk = FALSE,
  constant_values = NULL
)
}
\arguments{
\item{input_file}{the file name (including path) to be read}

\item{dbcon}{...}

\item{table_name}{...}

\item{columns}{...}

\item{drop_table}{...}

\item{auto_pk}{...}

\item{pk_fields}{...}

\item{build_pk}{...}

\item{constant_values}{...}
}
\value{
nothing
}
\description{
The dbTableFromFeather function reads the data from a Apache
Arrow table serialized in a Feather (Arrow IPC) file and copies it
to a table in a SQLite database. If table does not exist, it will
create it.
}
