% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromSchema.R
\name{dbTableFromSchema}
\alias{dbTableFromSchema}
\title{The dbTableFromSchema function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.}
\usage{
dbTableFromSchema(
  schema_file,
  dbcon,
  table_name,
  drop_table = FALSE,
  auto_pk = FALSE,
  build_pk = FALSE,
  constant_values = NULL
)
}
\arguments{
\item{schema_file}{...}

\item{dbcon}{...}

\item{table_name}{...}

\item{drop_table}{...}

\item{auto_pk}{...}

\item{build_pk}{...}

\item{constant_values}{...}
}
\value{
nothing
}
\description{
The dbTableFromSchema function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.
}
