% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromView.R
\name{dbTableFromView}
\alias{dbTableFromView}
\title{dbTableFromView create a table in a SQLite database from  a
view already present in the same database.}
\usage{
dbTableFromView(
  view_name,
  dbcon,
  table_name = NULL,
  drop_table = FALSE,
  pk_fields = NULL,
  build_pk = FALSE
)
}
\arguments{
\item{view_name}{...}

\item{dbcon}{database connection, as created by the dbConnect function.}

\item{table_name}{string, the name of the table.}

\item{drop_table}{...}

\item{pk_fields}{...}

\item{build_pk}{...}
}
\value{
nothing
}
\description{
dbTableFromView create a table in a SQLite database from  a
view already present in the same database.
}
