% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromXlsx.R
\name{dbTableFromXlsx}
\alias{dbTableFromXlsx}
\title{Create a table in a SQLite database from an Excel worksheet}
\usage{
dbTableFromXlsx(
  input_file,
  dbcon,
  table_name,
  sheet_name,
  first_row,
  cols_range,
  header = TRUE,
  id_quote_method = "DB_NAMES",
  col_names = NULL,
  col_types = NULL,
  col_import = NULL,
  drop_table = FALSE,
  auto_pk = FALSE,
  build_pk = FALSE,
  pk_fields = NULL,
  constant_values = NULL,
  ...
)
}
\arguments{
\item{input_file}{character, the file name (including path) to be read.}

\item{dbcon}{database connection, as created by the dbConnect function.}

\item{table_name}{character, the name of the table.}

\item{sheet_name}{character, the name of the worksheet containing the data
table.}

\item{first_row}{integer, the row number where the data table starts.
If present, it is the row number of the header row, otherwise it is
the row number of the first row of data.}

\item{cols_range}{integer, a numeric vector specifying which columns in
the worksheet to be read.}

\item{header}{logical, if \code{TRUE} the first row contains the fields'
names. If \code{FALSE}, the column names will be the column names of the
Excel worksheet (i.e. letters).}

\item{id_quote_method}{character, used to specify how to build the SQLite
columns' names using the fields' identifiers read from the input file.
For details see the description of the \code{quote_method} parameter of
the \code{\link[=format_column_names]{format_column_names()}} function. Defautls to \code{DB_NAMES}.}

\item{col_names}{character vector, names of the columuns in the input file.
Used to override the field names derived from the input file (using the
quote method selected by \code{id_quote_method}). Must be of the same length
of the number of columns in the input file. If \code{NULL} the column names
coming from the input file (after quoting) will be used.
Defaults to \code{NULL}.}

\item{col_types}{character vector of classes to be assumed for the columns
of the input file. Must be of the same length of the number of columns
in the input file. If not null, it will override the data types guessed
from the input file.
If \code{NULL} the data type inferred from the input files will be used.
Defaults to \code{NULL}.}

\item{col_import}{can be either:
\itemize{
\item a numeric vector (coherced to integers) with the columns' positions
in the input file that will be imported in the SQLite table;
\item a character vector with the columns' names to be imported. The names
are those in the input file (after quoting with \code{id_quote_method}),
if \code{col_names} is NULL, or those expressed in \code{col_names} vector.
Defaults to NULL, i.e. all columns will be imported.
}}

\item{drop_table}{logical, if \code{TRUE} the target table will be dropped (if
exists) and recreated before importing the data.  if \code{FALSE}, data from
input file will be appended to an existing table. Defaults to \code{FALSE}.}

\item{auto_pk}{logical, if \code{TRUE}, and \code{pk_fields} parameter is \code{NULL}, an
additional column named \code{SEQ} will be added to the table and it will be
defined to be \verb{INTEGER PRIMARY KEY} (i.e. in effect an alias for
\code{ROWID}). Defaults to \code{FALSE}.}

\item{build_pk}{logical, if \code{TRUE} creates a \verb{UNIQUE INDEX} named
\verb{<table_name>_PK} defined by the combination of fields specified
in the \code{pk_fields} parameter. It will be effective only if
\code{pk_fields} is not null. Defaults to \code{FALSE}.}

\item{pk_fields}{character vector, the list of the fields' names that
define the \verb{UNIQUE INDEX}. Defults to \code{NULL}.}

\item{constant_values}{a one row data frame whose columns will be added to
the table in the database. The additional table columns will be named
as the data frame columns, and the corresponding values will be associeted
to each record imported from the input file. It is useful to keep
track of additional information (e.g., the input file name, additional
context data not available in the data set, ...) when loading
the content of multiple input files in the same table. Defults to \code{NULL}.}

\item{...}{additional arguments passed to \code{\link[openxlsx2:wb_to_df]{openxlsx2::wb_to_df()}} function
used to read input data.}
}
\value{
integer, the number of records in \code{table_name} after reading data
from \code{input_file}.
}
\description{
The \code{dbTableFromXlsx()} function creates a table in a SQLite database from a
range of an Excel worksheet.

The \code{dbTableFromXlsx()} function reads the data from a range of
an Excel worksheet. If table does not exist, it will
create it.
}
\examples{
# Create a temporary database and load Excel data
library(RSQLite.toolkit)

# Set up database connection
dbcon <- dbConnect(RSQLite::SQLite(), file.path(tempdir(), "example.sqlite"))

# Get path to example data
data_path <- system.file("extdata", package = "RSQLite.toolkit")

# Check if Excel file exists (may not be available in all installations)
xlsx_file <- file.path(data_path, "stock_portfolio.xlsx")

fschema <- file_schema_xlsx(xlsx_file, sheet_name="all period", 
                            first_row=2, cols_range="A:S", header=TRUE,
                            id_quote_method="DB_NAMES", max_lines=10)

fschema[, c("col_names", "src_names")]

# Load Excel data from specific sheet and range
dbTableFromXlsx(
  input_file = xlsx_file,
  dbcon = dbcon,
  table_name = "PORTFOLIO_PERF",
  sheet_name = "all period",
  first_row = 2,
  cols_range = "A:S",
  drop_table = TRUE,
  col_import = c("ID", "Large_B_P", "Large_ROE", "Large_S_P",
                 "Annual_Return_7", "Excess_Return_8", "Systematic_Risk_9"),
  rows = c(2:23) 
)

# Check the imported data
dbListFields(dbcon, "PORTFOLIO_PERF")
head(dbGetQuery(dbcon, "SELECT * FROM PORTFOLIO_PERF"))

# Clean up
dbDisconnect(dbcon)

}
