% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error_handler}
\alias{error_handler}
\title{error_handler manage error messages for package}
\usage{
error_handler(err, fun, step)
}
\arguments{
\item{err}{character, error message}

\item{fun}{character, function name where error happened}

\item{step}{integer, code identifying the step in the function
where error happened.
For dbTableFrom... functions steps are:
\itemize{
\item 101,121: read file schema (DSV, Xlsx)
\item 102: handle col_names and col_types
\item 103: create empty table
\item 104: read data
\item 105: write data
\item 106: indexing
}}
}
\value{
nothing
}
\description{
error_handler manage error messages for package
}
