% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_helper.R
\name{feather_schema}
\alias{feather_schema}
\title{feather_schema returns a data frame with schema of the file}
\usage{
feather_schema(input.file)
}
\arguments{
\item{input.file}{the file name (including path) to be read}
}
\description{
feather_schema returns a data frame with schema of the file
}
\section{References}{

\href{https://stackoverflow.com/questions/66529055/how-to-read-column-names-and-metadata-from-feather-files-in-r-arrow}{Stackoverflow}
}

