% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{file_schema_feather}
\alias{file_schema_feather}
\title{file_schema_feather returns a data frame with the schema of a Feather file}
\usage{
file_schema_feather(input_file, id_quote_method = "DB_NAMES")
}
\arguments{
\item{input_file}{File name (including path) to be read}

\item{id_quote_method}{character, used to specify how to build the SQLite
columns' names using the fields' identifiers read from the input file.
For details see the description of the \code{quote_method} parameter of
the \code{\link[=format_column_names]{format_column_names()}} function. Defautls to \code{DB_NAMES}.}
}
\value{
a data frame with these columns:
\itemize{
\item \code{col_names}: columns' names, after applying the selected quote method;
\item \code{col_names_unquoted}: columns' names, unquoted; if \code{id_quote_method}
is set to \code{DB_NAMES} they will be the same as \code{col_names}; for other
quote methods they will be the unquoted versions of \code{col_names},that
is generally the same as \code{src_names} unless \code{src_names} contain the
quoting characters;
\item \code{col_types}: columns' R data types;
\item \code{sql_types}: columns' SQLite data types;
\item \code{src_names}: columns' names as they appear in the input file;
\item \code{src_types}: the Arrow's data type of each column.
}
}
\description{
Preview the table structure contained in a Feather file. This function
inspects the input file metadata to read the field identifiers' names and
data types, then converts them to the candidate data frame columns' names
and data types. The dataset contained in the input file is not read in to
memory, only meta-data are accessed.
}
\section{References}{

The implementation is based on this question on
\href{https://stackoverflow.com/questions/66529055/how-to-read-column-names-and-metadata-from-feather-files-in-r-arrow}{Stackoverflow}.
}

