% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_column_names}
\alias{format_column_names}
\title{Format column names for SQLite}
\usage{
format_column_names(
  x,
  quote_method = "DB_NAMES",
  unique_names = TRUE,
  encoding = ""
)
}
\arguments{
\item{x}{character vector with the identifiers' names to be quoted.}

\item{quote_method}{character, used to specify how to build the SQLite
columns' names from the identifiers passed through the \code{x}
parameter.
Supported values for \code{quote_method}:
\itemize{
\item \code{DB_NAMES} tries to build a valid SQLite column name:
a. substituting all characters, that are not letters or digits or
the \verb{_} character, with the \verb{_} character;
b. prefixing \code{N_} to all strings starting with a digit;
c. prefixing \code{F_} to all strings equal to any SQL92 keyword.
\item \code{SINGLE_QUOTES} encloses each string in single quotes.
\item \code{SQL_SERVER} encloses each string in square brackets.
\item \code{MYSQL} encloses each string in back ticks.
Defaults to \code{DB_NAMES}.
}}

\item{unique_names}{logical, checks for any duplicate name after
applying the selected quote methods. If duplicates exist, they
will be made unique by adding a postfix \verb{_[n]}, where \code{n} is
a progressive integer. Defaults to \code{TRUE}.}

\item{encoding}{character, encoding to be assumed for input strings.
It is used to re-encode the input in order to process it
to build column identifiers. Defaults to ‘""’ (for the encoding of
the current locale).}
}
\value{
A data frame containing the columns' identifiers in two formats:
\itemize{
\item \code{quoted}: the quoted names, as per the selected \code{quote_method};
\item \code{unquoted}: the cleaned names, without any quoting.
}
}
\description{
The \code{format_column_names()} function formats a vector of
strings to be used as columns' names for a table in a SQLite
database.
}
