% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R2SQL_types}
\alias{R2SQL_types}
\title{R2SQL_types returns a character vector with the names of SQLite data types
corresponding to the R classes passed through the \code{x} parameter.}
\usage{
R2SQL_types(x)
}
\arguments{
\item{x}{character, a vector containing the strings with the R class names.}
}
\value{
a character vector with the names of SQLite data types.
}
\description{
If any class is not recognized, it will be replaced with \code{TEXT} data type.
}
