% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{Xlsx_file_schema}
\alias{Xlsx_file_schema}
\title{Xlsx_file_schema returns a data frame with the schema of an Excel data table}
\usage{
Xlsx_file_schema(
  input_file,
  id_quote_method = "DB_NAMES",
  header = TRUE,
  sheet_name,
  first_row,
  cols_range,
  max_lines = 100,
  ...
)
}
\arguments{
\item{input_file}{File name (including path) to be read}

\item{id_quote_method}{Character, the quote method to be use to
form the column names from the field names read from the input
file.}

\item{header}{...}

\item{sheet_name}{...}

\item{first_row}{...}

\item{cols_range}{...}

\item{max_lines}{...}

\item{...}{parameters passed to openxlsx2::wb_to_df}
}
\value{
a data frame with the column names (\code{col_names})
and the Arrow's data type of each column (\code{src_types}).
}
\description{
Preview the table structure contained in a rectangula range of worksheet
of an Excel file. It can be used to select a subset of the columns to be
passed to the \code{dbTableFromXlsx} through the \code{col_names} parameter.
}
