% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{dbCopyTable}
\alias{dbCopyTable}
\title{dbCopyTable}
\usage{
dbCopyTable(
  db_file_src,
  db_file_tgt,
  table_name,
  drop_table = FALSE,
  pk_exist = TRUE
)
}
\arguments{
\item{db_file_src}{...}

\item{db_file_tgt}{...}

\item{table_name}{...}

\item{drop_table}{...}

\item{pk_exist}{...}
}
\description{
dbCopyTable
}
