% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{dbCreatePK}
\alias{dbCreatePK}
\title{Creates a unique index on a table in a SQLite database}
\usage{
dbCreatePK(dbcon, table_name, pk_fields, drop_index = FALSE)
}
\arguments{
\item{dbcon}{database connection, as created by the dbConnect function.}

\item{table_name}{character, the name of the table where the index
will be created.}

\item{pk_fields}{character vector, the list of the fields' names that
define the \verb{UNIQUE INDEX}.}

\item{drop_index}{logical, if \code{TRUE} the index named \verb{<table_name>_PK} will
be dropped (if exists) before recreating it. If \code{FALSE}, it will check
if an index with that name exists and eventually stops. Default to \code{FALSE}.}
}
\value{
nothing
}
\description{
The \code{dbCreatePK()} function creates a \verb{UNIQUE INDEX} named
\verb{<table_name>_PK} on the table specified by \code{table_name} in
the database connected by \code{dbcon}. The index is created on
the fields specified in the \code{pk_fields} argument.
}
