% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_helpers.R
\name{dbExecFile}
\alias{dbExecFile}
\title{Execute SQL statements from a text file}
\usage{
dbExecFile(input_file, dbcon, plist = NULL)
}
\arguments{
\item{input_file}{the file name (including path) containing the SQL
statements to be executed}

\item{dbcon}{database connection, as created by the dbConnect function.}

\item{plist}{a list with values to be binded to the parameters of
SQL statements. It should have the same length as the number of SQL
statements. If any of the statements do not require parameters,
the corresponding element of the list should be set to \code{NULL}.
If no statements require parameters, \code{plist} can be set to \code{NULL}.}
}
\value{
a list with the results returned by each statement executed.
}
\description{
The \code{dbExecFile()} function executes the SQL statements contained
in a text file.

This function reads the text in \code{input_file}, strips all comment lines
(i.e. all lines beginning with \verb{--} characters) and splits the SQL statements
assuming that they are separated by the \verb{;} character. The list of SQL
statements is then executed, one at a time; the results of each statement
are stored in a list with length equal to the number of statements.
}
