% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromDataFrame.R
\name{dbTableFromDataFrame}
\alias{dbTableFromDataFrame}
\title{The dbTableFromDataFrame function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.}
\usage{
dbTableFromDataFrame(
  df,
  dbcon,
  table_name,
  drop_table = FALSE,
  auto_pk = FALSE,
  pk_fields = NULL,
  build_pk = FALSE,
  constant_values = NULL
)
}
\arguments{
\item{df}{the data frame to be saved in SQLite table}

\item{dbcon}{...}

\item{table_name}{...}

\item{drop_table}{...}

\item{auto_pk}{...}

\item{pk_fields}{...}

\item{build_pk}{...}

\item{constant_values}{...}
}
\value{
a data frame with the structure of the table created in the
database
}
\description{
The dbTableFromDataFrame function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.
}
