% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromView.R
\name{dbTableFromView}
\alias{dbTableFromView}
\title{The dbTableFromView function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.}
\usage{
dbTableFromView(
  dbcon,
  view_name,
  table_name = NULL,
  pk_fields = NULL,
  build_pk = FALSE
)
}
\arguments{
\item{dbcon}{...}

\item{view_name}{...}

\item{table_name}{...}

\item{pk_fields}{...}

\item{build_pk}{...}
}
\value{
nothing
}
\description{
The dbTableFromView function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.
}
