% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbTableFromXlsx.R
\name{dbTableFromXlsx}
\alias{dbTableFromXlsx}
\title{The dbTableFromXlsx function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.}
\usage{
dbTableFromXlsx(
  input_file,
  sheet_name,
  first_row,
  header = TRUE,
  dbcon,
  table_name,
  drop_table = FALSE,
  auto_pk = FALSE,
  pk_fields = NULL,
  build_pk = FALSE,
  constant_values = NULL,
  ...
)
}
\arguments{
\item{input_file}{the file name (including path) to be read}

\item{sheet_name}{...}

\item{first_row}{...}

\item{header}{...}

\item{dbcon}{...}

\item{table_name}{...}

\item{drop_table}{...}

\item{auto_pk}{...}

\item{pk_fields}{...}

\item{build_pk}{...}

\item{constant_values}{...}

\item{...}{...}
}
\value{
nothing
}
\description{
The dbTableFromXlsx function reads the data from a rectangula region
of a sheet in an Excel file and copies it to a table in a SQLite
database. If table does not exist, it will create it.
}
