% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{Feather_file_schema}
\alias{Feather_file_schema}
\title{Feather_file_schema returns a data frame with the schema of a Feather file}
\usage{
Feather_file_schema(input_file, id_quote_method = "DB_NAMES")
}
\arguments{
\item{input_file}{File name (including path) to be read}

\item{id_quote_method}{Character, the quote method to be use to
form the column names from the field names read from the input
file.}
}
\value{
a data frame with the column names (\code{col_names})
and the Arrow's data type of each column (\code{src_types}).
}
\description{
Preview the table structure contained in a Feather file. It can be used
to select a subset of the columns to be passed to the \code{dbTableFromFeather}
through the \code{columns} parameter.
}
\section{References}{

The implementation is based on this question on
\href{https://stackoverflow.com/questions/66529055/how-to-read-column-names-and-metadata-from-feather-files-in-r-arrow}{Stackoverflow}.
}

