% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{feather_schema}
\alias{feather_schema}
\title{feather_schema returns a data frame with schema of the Feather file}
\usage{
feather_schema(input.file)
}
\arguments{
\item{input.file}{the file name (including path) to be read}
}
\value{
a data frame with the column names (\code{col_names})
and the Arrow data type of each column (\code{type_names}).
}
\description{
This helper function can be used to preview the table structure contained
in a Feather file in order to select a subset of the columns to be
passed to the \code{dbTableFromFeather} through the \code{columns} parameter.
}
\section{References}{

The implementation is based on this questioon on \href{https://stackoverflow.com/questions/66529055/how-to-read-column-names-and-metadata-from-feather-files-in-r-arrow}{Stackoverflow}.
}

